export and unify WorkletEventHandler detection
